pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[all]>=0.115.12",
    "google-generativeai>=0.8.5",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.41",
]
